[refactor] 경로
