+++ src/main/java/com/example/coffee/common/jwt/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package com.example.coffee.common.config;
+package com.example.coffee.common.jwt;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
@@ -13,7 +13,7 @@ import java.util.Date;
 public class JwtTokenProvider {
 
     private final SecretKey secretKey; // 동적으로 생성된 키
-    private final long validityInMilliseconds = 3600000; // 1시간 (예제 값)
+    private final long validityInMilliseconds = 360000; // 1시간 (예제 값)
 
     // 키 생성자
     public JwtTokenProvider() {
@@ -22,7 +22,7 @@ public class JwtTokenProvider {
 
     // JWT 생성
     public String createToken(long id) {
-        Claims claims = Jwts.claims().setSubject(String.valueOf(id)); // 이메일 정보 저장
+        Claims claims = Jwts.claims().setSubject(String.valueOf(id));
         Date now = new Date();
         Date validity = new Date(now.getTime() + validityInMilliseconds);
